more secure password system
upon reflection, I'd rather not require the user to store their FTP
password in plaintext. that's insecure and I know it. so now, the user
will have to input their FTP password manually whenever they publish to
web, which shouldn't be so often that it's a pain.
in short: much more secure, slightly less convenient.